@@ -1,13 +1,14 @@
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+using ObjectsExchange.Data;
 namespace ObjectsExchange;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.AspNetCore.Mvc.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Identity.Web;
-using Microsoft.Identity.Web.UI;
 using ObjectsExchange.Data;
 using Westwind.AspNetCore.Markdown;
 
@@ -16,17 +17,35 @@ public class Program
     public static void Main(string[] args)
     {
         var builder = WebApplication.CreateBuilder(args);
+        var connectionString = builder.Configuration.GetConnectionString("DefaultConnection") ?? throw new InvalidOperationException("Connection string 'DefaultConnection' not found."); ;
+
 
         builder.Services.AddDbContext<ObjectsExchangeDbContext>(options =>
-                    options.UseSqlite(builder.Configuration.GetConnectionString("SqliteConnection")));
+                    options.UseNpgsql(connectionString));
+
+        builder.Services.AddDatabaseDeveloperPageExceptionFilter();
+        builder.Services.AddLocalization();
+
+        builder.Services.AddDefaultIdentity<IdentityUser>(options => options.SignIn.RequireConfirmedAccount = true)
+            .AddEntityFrameworkStores<ObjectsExchangeDbContext>();
 
         builder.Services.Configure<ObjectsExchange.Services.ApiConfig>(
                    builder.Configuration.GetSection(nameof(ObjectsExchange.Services.ApiConfig)));
 
+        // add Microsoft account
+        var autenficationBuilder = builder.Services.AddAuthentication();
 
-        // Add services to the container.
-        builder.Services.AddAuthentication(OpenIdConnectDefaults.AuthenticationScheme)
-            .AddMicrosoftIdentityWebApp(builder.Configuration.GetSection("AzureAd"));
+        var microsoftClientId = builder.Configuration["Authentication:Microsoft:ClientId"];
+        var microsoftClientSecret = builder.Configuration["Authentication:Microsoft:ClientSecret"];
+        if (!(string.IsNullOrWhiteSpace(microsoftClientId) || string.IsNullOrWhiteSpace(microsoftClientSecret)))
+        {
+            autenficationBuilder.AddMicrosoftAccount(microsoftOptions =>
+            {
+                microsoftOptions.ClientId = microsoftClientId;
+                microsoftOptions.ClientSecret = microsoftClientSecret;
+            });
+        }
+
 
         builder.Services.AddAuthorization(options =>
         {
@@ -35,8 +54,18 @@ public class Program
         });
         builder.Services.AddMarkdown();
 
-        builder.Services.AddRazorPages().AddMicrosoftIdentityUI().AddApplicationPart(typeof(MarkdownPageProcessorMiddleware).Assembly);
+        builder.Services.AddRazorPages()
+                        .AddViewLocalization()
+                        .AddDataAnnotationsLocalization()
+                        .AddApplicationPart(typeof(MarkdownPageProcessorMiddleware).Assembly);
         builder.Services.AddScoped<ObjectsExchange.Services.ClientManager>();
+
+        builder.Services.AddAuthentication();
+        builder.Services.Configure<ForwardedHeadersOptions>(options =>
+        {
+            options.ForwardedHeaders =
+                ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto;
+        });
 
         var app = builder.Build();
 
@@ -49,6 +78,7 @@ public class Program
             app.UseHttpsRedirection();
         }
 
+        app.UseHttpsRedirection();
         app.UseMarkdown();
         app.UseStaticFiles();
         app.UseRouting();
@@ -58,6 +88,13 @@ public class Program
         });
         app.UseAuthentication();
         app.UseAuthorization();
+
+        app.UseRequestLocalization(
+            new RequestLocalizationOptions() { ApplyCurrentCultureToResponseHeaders = true }
+                .AddSupportedCultures(new[] { "en-US", "uk-UA" })
+                .AddSupportedUICultures(new[] { "en-US", "uk-UA" })
+                .SetDefaultCulture("uk-UA")
+                );
 
         app.MapRazorPages();
         app.MapControllers();
