@@ -1,7 +1,5 @@
 ﻿    using Sabatex.Core;
-#if NET5_0_OR_GREATER
     using System.ComponentModel.DataAnnotations;
-#endif
     using System;
 
 namespace Sabatex.ObjectsExchange.Models
@@ -9,7 +7,7 @@ namespace Sabatex.ObjectsExchange.Models
     /// <summary>
     /// Object for exchange
     /// </summary>
-    public class ObjectExchange : IEntityBase
+    public class ObjectExchange : IEntityBase<long>
     {
         /// <summary>
         /// primary key
@@ -26,17 +24,13 @@ namespace Sabatex.ObjectsExchange.Models
         /// <summary>
         /// internal unique client object id  
         /// </summary>
-#if NET6_0_OR_GREATER
         [Required]
         [MaxLength(50)]
-#endif
         public string ObjectId { get; set; } = default!;
         /// <summary>
         /// client object type
         /// </summary>
-#if NET6_0_OR_GREATER
         [MaxLength(50)]
-#endif
         public string ObjectType { get; set; } = default!;
         /// <summary>
         /// внутрішня позначка часу створення обєкта
@@ -50,6 +44,6 @@ namespace Sabatex.ObjectsExchange.Models
         /// serialized object (json,xml,csv...)
         /// </summary>
         public string ObjectAsText { get; set; } = default!;
-        string IEntityBase.KeyAsString() => Id.ToString();
+        //string IEntityBase.KeyAsString() => Id.ToString();
     }
 }
